requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
pydantic
sqlalchemy
pytest
